feat: externalize config loading in its package
