+++ packages/config/build.config.ts
@@ -3,17 +3,9 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig({
 	entries: [
 		'src/index',
-		'src/vite',
-		'src/config',
-		'src/vue',
 	],
 	clean: true,
 	declaration: true,
-	externals: [
-		'vite',
-		'vue',
-		'@hybridly/progress-plugin',
-	],
 	rollup: {
 		emitCJS: true,
 	},
