+++ packages/core/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@hybridly/core",
   "version": "0.0.1-alpha.24",
-  "description": "A solution to develop server-driven, client-rendered applications",
+  "description": "Core functionality of Hybridly",
   "keywords": [
     "hybridly",
     "inertiajs"
