+++ packages/hybridly/src/index.ts
@@ -1,2 +1,1 @@
 export * from '@hybridly/core'
-export * from './config'
