+++ packages/vite/build.config.ts
@@ -10,6 +10,7 @@ export default defineBuildConfig({
 		'vite',
 		'debug',
 		'hybridly',
+		'@hybridly/config',
 	],
 	rollup: {
 		emitCJS: true,
