+++ packages/vite/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@hybridly/vite",
   "version": "0.0.1-alpha.24",
-  "description": "A solution to develop server-driven, client-rendered applications",
+  "description": "Vite plugin for Hybridly",
   "keywords": [
     "hybridly",
     "inertiajs",
