+++ packages/vite/src/config/index.ts
@@ -1,7 +1,6 @@
 import path from 'node:path'
 import type { Plugin } from 'vite'
-import type { ResolvedHybridlyConfig } from 'hybridly'
-import { loadHybridlyConfig } from 'hybridly'
+import type { ResolvedHybridlyConfig } from '@hybridly/config'
 import { CONFIG_PLUGIN_NAME, CONFIG_VIRTUAL_MODULE_ID, RESOLVED_CONFIG_VIRTUAL_MODULE_ID, ROUTING_VIRTUAL_MODULE_ID } from '../constants'
 import type { ViteOptions } from '../types'
 
@@ -45,8 +44,6 @@ export default (options: ViteOptions, config: ResolvedHybridlyConfig): Plugin =>
 		},
 		async load(id) {
 			if (id === RESOLVED_CONFIG_VIRTUAL_MODULE_ID) {
-				const config = await loadHybridlyConfig()
-
 				return `
 					import { initializeHybridly as init } from 'hybridly/vue'
 					import '${ROUTING_VIRTUAL_MODULE_ID}'
