+++ packages/vue/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@hybridly/vue",
   "version": "0.0.1-alpha.24",
-  "description": "A solution to develop server-driven, client-rendered applications",
+  "description": "Vue adapter for Hybridly",
   "keywords": [
     "hybridly",
     "inertiajs",
@@ -45,6 +45,7 @@
     "@hybridly/core": "workspace:*",
     "@hybridly/progress-plugin": "workspace:*",
     "@hybridly/utils": "workspace:*",
+    "@hybridly/config": "workspace:*",
     "@vue/devtools-api": "^6.5.0",
     "defu": "^6.1.2",
     "lodash.isequal": "^4.5.0",
